Screenshot tool functions have been organized into the correct file
