+++ TutorialMaker/TutorialMaker.py
@@ -174,42 +174,12 @@ class TutorialMakerLogic(ScriptedLoadableModuleLogic):
         pass
 
     def ExportScreenshots(self):
-        self.saveScreenshotMetadata(0)
+        screenshot = utils.ScreenshotTools()
+        screenshot.saveScreenshotMetadata(0)
         pass
 
     # TODO Transfer these functions to the Utils class
-    def saveScreenshotMetadata(self, index):
-        self.saveAllWidgetsData(os.path.dirname(slicer.util.modulePath("TutorialMaker")) + "/Outputs/output_tutorial."+str(index)+".json")
-        self.saveScreenshot(os.path.dirname(slicer.util.modulePath("TutorialMaker")) + "/Outputs/output_tutorial."+str(index)+".png")
-        pass
-
-
-    def getPixmap(self):
-        mw = slicer.util.mainWindow()
-        screen = slicer.app.screens()[0]
-        pixmap = screen.grabWindow(mw.winId())
-
-        #return a qt object: QPixmap
-        return pixmap
     
-    def saveScreenshot(self, filename):
-        self.getPixmap().save(filename, "PNG")
-        pass
-
-    def saveAllWidgetsData(self, filename):
-        util = utils.util()
-        data = {}
-        widgets = util.getOnScreenWidgets()
-        for index in range(widgets.__sizeof__()):
-            try:
-                data[index] = {"name": widgets[index].name, "path": util.uniqueWidgetPath(widgets[index]), "text": widgets[index].text, "position": widgets[index].getGlobalPos(), "size": widgets[index].getSize()}
-            except:
-                pass
-        import json
-        with open(filename, 'w', encoding='utf-8') as f:
-            json.dump(data, f, ensure_ascii=False, indent=4)
-        pass
-
 #
 # TutorialMakerTest
 #
